+++ src/sentry/issues/priority.py
@@ -1,30 +1,19 @@
 from __future__ import annotations
 
 import logging
-from enum import Enum, IntEnum
-from typing import TYPE_CHECKING, Type
+from enum import Enum
+from typing import TYPE_CHECKING
 
 from sentry import features
-from sentry.issues.grouptype import (
-    GroupCategory,
-    GroupType,
-    PerformanceP95EndpointRegressionGroupType,
-    ProfileFunctionRegressionType,
-)
 from sentry.models.activity import Activity
 from sentry.models.grouphistory import GroupHistory, GroupHistoryStatus, record_group_history
 from sentry.models.user import User
 from sentry.services.hybrid_cloud.user.model import RpcUser
 from sentry.types.activity import ActivityType
+from sentry.types.group import PriorityLevel
 
 if TYPE_CHECKING:
     from sentry.models.group import Group
-
-
-class PriorityLevel(IntEnum):
-    LOW = 25
-    MEDIUM = 50
-    HIGH = 75
 
 
 PRIORITY_LEVEL_TO_STR: dict[int, str] = {
@@ -136,27 +125,3 @@ def auto_update_priority(group: Group, reason: PriorityChangeReason) -> None:
 
     if new_priority is not None:
         update_priority(group, new_priority, reason)
-
-
-def get_default_priority_for_group_type(group_type: Type[GroupType], level: str) -> PriorityLevel:
-    if group_type.category in [GroupCategory.REPLAY.value, GroupCategory.FEEDBACK.value]:
-        return PriorityLevel.MEDIUM
-
-    if group_type.category == GroupCategory.CRON.value:
-        if level == "warning":
-            return PriorityLevel.MEDIUM
-
-        return PriorityLevel.HIGH
-
-    # Profiling issues should be treated the same as Performance issues since they are merging
-    if group_type.category in [GroupCategory.PERFORMANCE.value, GroupCategory.PROFILE.value]:
-        # Statistical detectors are medium priority
-        if group_type.id in [
-            ProfileFunctionRegressionType.type_id,
-            PerformanceP95EndpointRegressionGroupType.type_id,
-        ]:
-            return PriorityLevel.MEDIUM
-        return PriorityLevel.LOW
-
-    # All other issues are the default medium priority
-    return PriorityLevel.MEDIUM
