+++ src/sentry/types/group.py
@@ -1,3 +1,4 @@
+from enum import IntEnum
 from typing import Mapping
 
 
@@ -65,3 +66,9 @@ GROUP_SUBSTATUS_TO_GROUP_HISTORY_STATUS = {
     GroupSubStatus.FOREVER: "archived_forever",
     GroupSubStatus.UNTIL_CONDITION_MET: "archived_until_condition_met",
 }
+
+
+class PriorityLevel(IntEnum):
+    LOW = 25
+    MEDIUM = 50
+    HIGH = 75
